# Slightly speedup ASTInterpreter::initArguments