--- a/src/codegen/irgen/hooks.cpp
+++ b/src/codegen/irgen/hooks.cpp
@@ -45,7 +45,8 @@
 namespace pyston {
 
 // TODO terrible place for these!
-ParamNames::ParamNames(AST* ast, InternedStringPool& pool) : takes_param_names(true) {
+ParamNames::ParamNames(AST* ast, InternedStringPool& pool)
+    : takes_param_names(true), vararg_name(NULL), kwarg_name(NULL) {
     if (ast->type == AST_TYPE::Module || ast->type == AST_TYPE::ClassDef || ast->type == AST_TYPE::Expression
         || ast->type == AST_TYPE::Suite) {
         kwarg = "";
@@ -56,22 +57,30 @@ ParamNames::ParamNames(AST* ast, InternedStringPool& pool) : takes_param_names(t
         for (int i = 0; i < arguments->args.size(); i++) {
             AST_expr* arg = arguments->args[i];
             if (arg->type == AST_TYPE::Name) {
-                args.push_back(ast_cast<AST_Name>(arg)->id.s());
+                AST_Name* name = ast_cast<AST_Name>(arg);
+                arg_names.push_back(name);
+                args.push_back(name->id.s());
             } else {
                 InternedString dot_arg_name = pool.get("." + std::to_string(i));
+                arg_names.push_back(new AST_Name(dot_arg_name, AST_TYPE::Param, arg->lineno, arg->col_offset));
                 args.push_back(dot_arg_name.s());
             }
         }
 
         vararg = arguments->vararg.s();
+        if (vararg.size())
+            vararg_name = new AST_Name(pool.get(vararg), AST_TYPE::Param, arguments->lineno, arguments->col_offset);
+
         kwarg = arguments->kwarg.s();
+        if (kwarg.size())
+            kwarg_name = new AST_Name(pool.get(kwarg), AST_TYPE::Param, arguments->lineno, arguments->col_offset);
     } else {
         RELEASE_ASSERT(0, "%d", ast->type);
     }
 }
 
 ParamNames::ParamNames(const std::vector<llvm::StringRef>& args, llvm::StringRef vararg, llvm::StringRef kwarg)
-    : takes_param_names(true) {
+    : takes_param_names(true), vararg_name(NULL), kwarg_name(NULL) {
     this->args = args;
     this->vararg = vararg;
     this->kwarg = kwarg;