This function can get quite hot because now with the baseline jit we execute it much more often.
I'm not really excited about adding the ```InternedString``` fields to ```ParamNames``` but I don't know where to add them else.
```
                           e648044ce2266a3fc6:  b8f5647281593fc5ed:
       django_template.py             3.5s (4)             3.4s (4)  -0.7%
            pyxl_bench.py             3.3s (4)             3.3s (4)  -0.6%
sqlalchemy_imperative2.py             4.1s (4)             4.1s (4)  -0.7%
                  geomean                 3.6s                 3.6s  -0.7%
```