@undingen 2015-07-19
I added another optimization (cache lookup type) and changed the ```ParamNames``` to contain ```AST_Name``` nodes. (Could be nicer any ideas?)

perf is now:
```

       django_template.py             3.5s (4)             3.4s (4)  -1.1%
            pyxl_bench.py             3.3s (4)             3.3s (4)  -1.1%
sqlalchemy_imperative2.py             4.1s (4)             4.1s (4)  -1.2%
                  geomean                 3.6s                 3.6s  -1.1%

```